+++ src/sections/Route.enum.ts
@@ -21,9 +21,10 @@ export enum Route {
   EDIT_COLLECTION = '/collections/:collectionId/edit',
   EDIT_FEATURED_ITEMS = '/collections/:collectionId/edit-featured-items',
   COLLECTION_TEMPLATES = '/:collectionId/templates',
+  GUESTBOOKS = '/:collectionId/guestbooks',
+  GUESTBOOKS_CREATE = '/:collectionId/guestbooks/create',
   TEMPLATES_CREATE = '/:collectionId/templates/create',
-  TEMPLATES_EDIT_METADATA = '/:collectionId/templates/:templateId/edit/metadata',
-  TEMPLATES_EDIT_TERMS = '/:collectionId/templates/:templateId/edit/terms',
+  TEMPLATES_EDIT = '/templates/edit',
   FEATURED_ITEM = '/featured-item/:parentCollectionId/:featuredItemId',
   NOT_FOUND_PAGE = '/404',
   AUTH_CALLBACK = '/auth-callback',
@@ -39,11 +40,21 @@ export const RouteWithParams = {
   EDIT_COLLECTION: (collectionId: string) => `/collections/${collectionId}/edit`,
   EDIT_FEATURED_ITEMS: (collectionId: string) => `/collections/${collectionId}/edit-featured-items`,
   COLLECTION_TEMPLATES: (collectionId: string) => `/${collectionId}/templates`,
+  GUESTBOOKS: (collectionId: string) => `/${collectionId}/guestbooks`,
+  GUESTBOOKS_CREATE: (collectionId: string) => `/${collectionId}/guestbooks/create`,
   TEMPLATES_CREATE: (collectionId: string) => `/${collectionId}/templates/create`,
-  TEMPLATES_EDIT_METADATA: (collectionId: string, templateId: number | string) =>
-    `/${collectionId}/templates/${templateId}/edit/metadata`,
-  TEMPLATES_EDIT_TERMS: (collectionId: string, templateId: number | string) =>
-    `/${collectionId}/templates/${templateId}/edit/terms`,
+  TEMPLATES_EDIT: (
+    collectionId: string,
+    templateId: number | string,
+    editMode: TemplateEditMode
+  ) => {
+    const searchParams = new URLSearchParams({
+      [QueryParamKey.ID]: templateId.toString(),
+      [QueryParamKey.OWNER_ID]: collectionId,
+      [QueryParamKey.EDIT_MODE]: editMode
+    })
+    return `${Route.TEMPLATES_EDIT}?${searchParams.toString()}`
+  },
   EDIT_FILE_METADATA: (
     datasetPersistentId: string,
     datasetVersion: string,
@@ -89,9 +100,17 @@ export enum QueryParamKey {
   COLLECTION_ID = 'collectionId',
   TAB = 'tab',
   FILE_ID = 'id',
+  ID = 'id',
+  OWNER_ID = 'ownerId',
+  EDIT_MODE = 'editMode',
   DATASET_VERSION = 'datasetVersion',
   REFERRER = 'referrer',
   AUTH_STATE = 'state',
   VALID_TOKEN_BUT_NOT_LINKED_ACCOUNT = 'validTokenButNotLinkedAccount',
   TOOL_TYPE = 'toolType'
+}
+
+export enum TemplateEditMode {
+  METADATA = 'METADATA',
+  LICENSE = 'LICENSE'
 }