+++ src/sections/templates/create-template/CreateTemplate.tsx
@@ -1,11 +1,13 @@
+import { useEffect } from 'react'
 import { useTranslation } from 'react-i18next'
-import { BreadcrumbsGenerator } from '../../shared/hierarchy/BreadcrumbsGenerator'
 import { CollectionRepository } from '@/collection/domain/repositories/CollectionRepository'
 import { TemplateRepository } from '@/templates/domain/repositories/TemplateRepository'
 import { MetadataBlockInfoRepository } from '@/metadata-block-info/domain/repositories/MetadataBlockInfoRepository'
 import { useCollection } from '@/sections/collection/useCollection'
 import { NotFoundPage } from '@/sections/not-found-page/NotFoundPage'
 import { TemplateMetadataForm } from '@/sections/shared/form/TemplateMetadataForm/TemplateMetadataForm'
+import { BreadcrumbsGenerator } from '@/sections/shared/hierarchy/BreadcrumbsGenerator'
+import { useLoading } from '@/shared/contexts/loading/LoadingContext'
 import { CreateTemplateSkeleton } from './CreateTemplateSkeleton'
 import styles from './CreateTemplate.module.scss'
 
@@ -23,12 +25,17 @@ export const CreateTemplate = ({
   metadataBlockInfoRepository
 }: CreateTemplateProps) => {
   const { t } = useTranslation('datasetTemplates')
+  const { setIsLoading } = useLoading()
   const { collection, isLoading: isLoadingCollection } = useCollection(
     collectionRepository,
     collectionId
   )
 
   const isLoadingData = isLoadingCollection
+
+  useEffect(() => {
+    setIsLoading(isLoadingData)
+  }, [isLoadingData, setIsLoading])
 
   if (!isLoadingCollection && !collection) {
     return <NotFoundPage dvObjectNotFoundType="collection" />
@@ -40,11 +47,16 @@ export const CreateTemplate = ({
 
   return (
     <section className={styles.container}>
-      <BreadcrumbsGenerator hierarchy={collection.hierarchy}></BreadcrumbsGenerator>
+      <BreadcrumbsGenerator
+        hierarchy={collection.hierarchy}
+        withActionItem
+        actionItemText={t('createTemplate.pageTitle')}
+      />
       <header className={styles.header}>
         <h1>{t('createTemplate.pageTitle')}</h1>
       </header>
       <TemplateMetadataForm
+        mode="create"
         collectionId={collectionId}
         metadataBlockInfoRepository={metadataBlockInfoRepository}
         templateRepository={templateRepository}