+++ src/stories/templates/TemplateMockRepository.ts
@@ -1,6 +1,9 @@
 import { TemplateInfo } from '@/templates/domain/models/TemplateInfo'
+import { UpdateTemplateMetadataInfo } from '@/templates/domain/models/UpdateTemplateMetadataInfo'
+import { UpdateTemplateLicenseTermsInfo } from '@/templates/domain/models/UpdateTemplateLicenseTermsInfo'
 import { TemplateRepository } from '@/templates/domain/repositories/TemplateRepository'
 import { Template } from '@/templates/domain/models/Template'
+import { TermsOfAccess } from '@/dataset/domain/models/Dataset'
 import { TemplateMother } from '@tests/component/sections/templates/TemplateMother'
 import { FakerHelper } from '@tests/component/shared/FakerHelper'
 
@@ -30,6 +33,53 @@ export class TemplateMockRepository implements TemplateRepository {
   }
 
   deleteTemplate(_templateId: number): Promise<void> {
+    return new Promise((resolve) => {
+      setTimeout(() => {
+        resolve()
+      }, FakerHelper.loadingTimout())
+    })
+  }
+
+  setTemplateAsDefault(_templateId: number, _collectionIdOrAlias: number | string): Promise<void> {
+    return new Promise((resolve) => {
+      setTimeout(() => {
+        resolve()
+      }, FakerHelper.loadingTimout())
+    })
+  }
+
+  unsetTemplateAsDefault(_collectionIdOrAlias: number | string): Promise<void> {
+    return new Promise((resolve) => {
+      setTimeout(() => {
+        resolve()
+      }, FakerHelper.loadingTimout())
+    })
+  }
+
+  updateTemplateMetadata(
+    _templateId: number,
+    _payload: UpdateTemplateMetadataInfo,
+    _replace?: boolean
+  ): Promise<void> {
+    return new Promise((resolve) => {
+      setTimeout(() => {
+        resolve()
+      }, FakerHelper.loadingTimout())
+    })
+  }
+
+  updateTemplateLicenseTerms(
+    _templateId: number,
+    _payload: UpdateTemplateLicenseTermsInfo
+  ): Promise<void> {
+    return new Promise((resolve) => {
+      setTimeout(() => {
+        resolve()
+      }, FakerHelper.loadingTimout())
+    })
+  }
+
+  updateTemplateTermsOfAccess(_templateId: number, _termsOfAccess: TermsOfAccess): Promise<void> {
     return new Promise((resolve) => {
       setTimeout(() => {
         resolve()