+++ src/templates/domain/useCases/unsetTemplateAsDefault.ts
@@ -1,0 +1,8 @@
+import { TemplateRepository } from '../repositories/TemplateRepository'
+
+export function unsetTemplateAsDefault(
+  templateRepository: TemplateRepository,
+  collectionIdOrAlias: number | string
+): Promise<void> {
+  return templateRepository.unsetTemplateAsDefault(collectionIdOrAlias)
+}